+++ m3_for_arty_a7/lcd/dis_plate.h
@@ -1,6 +1,6 @@
 #include "xil_types.h"
   
 
-void show_plate_num(u32 *p, u8 position);
+void show_plate_num(u8 *p, u8 position);
 void show_plate_chinese(u8 chinese_dis);
 void show_plate_english(u8 english_dis);
